+++ Changelog.md
@@ -16,4 +16,13 @@ Contains notable changes there were added, fixed, or removed in each release.
 
 ### Changes
 
-* ADS8689IPWR driver
+* ADS8689IPWR driver created and implemented
+* RedundantADC driver created and implemented
+* HIB driver created and implemented
+* Updated README.md
+* Created targets for HIB, RedundantADC testing, and ADS88689IPWR testing
+* Improved dependency order
+* Verified ADC results
+* Added error checking for ADCs
+* Added logger statements for easier debugging
+* Documentation added for each class as well as each target